fix: rearrange videos file to avoid menu junk
Rather than trying to fix the items in the menu which relate to
all sorts of things, it's better to move videos into a sub folder.

We shouldn't be putting standalone files into /developer, because
it is a navigation-level folder in the tree.

sample_template doesn't need work because it is a template for
use elsewhere; it may not be needed but not doing harm sitting
there.

Fixes #160.