docs(core): add 17.0 docs
The files json-schema.md and building.md are local to this repo. The
remainder are generated from keyman repo, keyman_core_api.h by
api-header-extractor, and manually copied in from core/build/doc.

In the future, we need to move json-schema.md and building.md into the
keyman repo as well, and generate the whole folder.

Also in the future we need to eliminate refs to Keyman Core Web and move
this entire structure up a level, as it's just "Keyman Core" now, not
"Keyman Core Desktop".