Update packages.md
The wording here is unclear. "Step 1" then "Step 2" in the headings implies you need to do #1 then do #2. But these are two alternate ways to share a kmp file. You can share it directly, or via a cloud service without putting it on a website, or you can put it on a website without sharing directly, or you can do both.  My wording tries to express this.