fix: support __index.txt to custom sort help page index
If a file __index.txt exists in a folder, it should list the .php and
.md files in the desired order for display in the left hand navigation
menu. A file prefixed by a hyphen will be excluded from the index, and
other files not in the file will be listed in their default alphabetical
order afterwards.

The example __index.txt files will be added to the keyman repository for
help deployment.

Test-bot: skip